split the message spec with ability to specify spec version to test and import that version
